requirements.txt
beautifulsoup4==4.12.2
Django==3.1.14
Django==4.2.2
djangorestframework==3.11.2
djangorestframework==3.14.0
langchain==0.0.201
pandas==2.0.2
pydantic==1.10.8
Requests==2.31.0
sampling==0.0.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
